Iteration18 (#625)

+++ elenasrc3/engine/windows/winroutines.cpp
@@ -62,7 +62,7 @@ uintptr_t SystemRoutineProvider :: ExpandPerm(void* allocPtr, size_t newSize)
    return !r ? 0 : (uintptr_t)allocPtr;
 }
 
-void* SystemRoutineProvider :: CreateThread(int tt_index, int flags, void* threadProc)
+void* SystemRoutineProvider :: CreateThread(size_t tt_index, int flags, void* threadProc)
 {
    return ::CreateThread(
       nullptr,                            // default security attributes
@@ -194,7 +194,7 @@ long long SystemRoutineProvider :: GenerateSeed()
    return seed;
 }
 
-void SystemRoutineProvider :: InitMTASignals(SystemEnv* env, int index)
+void SystemRoutineProvider :: InitMTASignals(SystemEnv* env, size_t index)
 {
    env->th_table->slots[index].content->tt_sync_event = ::CreateEvent(0, -1, 0, 0);
    env->th_table->slots[index].content->tt_flags = 0;
